+++ packages/react-components/source/react/library/input/Input.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Icon from '../icon';
+import Button from '../button';
 
 /**
  * This corresponds to a set of native input types plus 'multiline',
@@ -35,8 +36,14 @@ const propTypes = {
   error: PropTypes.oneOfType([PropTypes.string, PropTypes.bool]),
   /** Optional icon rendered before input area */
   icon: PropTypes.string,
-  /** Optional icon rendered after input area */
+  /** @deprecated Optional icon rendered after input area */
   trailingIcon: PropTypes.string,
+  /** Icon for rendered trailing button */
+  trailingButtonIcon: PropTypes.string,
+  /** Text for rendered trailing button. Can be used with or without trailingButtonIcon */
+  trailingButtonText: PropTypes.string,
+  /** Additional props for the trailing Button */
+  trailingButtonProps: PropTypes.shape({}),
   /** Optional additional className */
   className: PropTypes.string,
   /** Optional inline styles */
@@ -45,6 +52,8 @@ const propTypes = {
   inputRef: PropTypes.func,
   /** Change handler. Passed in order: new value, original event. Additionally, other event handlers and and props are propagated to the inner input element for use as needed */
   onChange: PropTypes.func,
+  /** Function for trailing button click */
+  onClickTrailingButton: PropTypes.func,
 };
 
 const defaultProps = {
@@ -56,10 +65,14 @@ const defaultProps = {
   error: false,
   icon: null,
   trailingIcon: null,
+  trailingButtonIcon: null,
+  trailingButtonText: '',
+  trailingButtonProps: {},
   style: {},
   className: '',
   inputRef() {},
   onChange() {},
+  onClickTrailingButton() {},
 };
 
 /**
@@ -81,34 +94,59 @@ const Input = ({
   error,
   icon,
   trailingIcon,
+  trailingButtonIcon,
+  trailingButtonText,
+  trailingButtonProps,
   className,
   style,
   inputRef,
   onChange,
+  onClickTrailingButton,
   ...otherProps
 }) => {
   const isMultiline = type === 'multiline';
 
   const Element = isMultiline ? 'textarea' : 'input';
 
+  const showTrailingButton = !!trailingButtonIcon || !!trailingButtonText;
+
+  const lIcon = (
+    <Icon
+      className="rc-input-icon leading"
+      width="16px"
+      height="16px"
+      type={icon}
+    />
+  );
+
+  /** trailingIcon is deprecated */
+  const tIcon = (
+    <Icon
+      className={`rc-input-icon trailing ${showTrailingButton &&
+        'with-trailing-button'}`}
+      width="16px"
+      height="16px"
+      type={trailingIcon}
+    />
+  );
+
+  const trailingButton = (
+    <Button
+      className="rc-input-icon rc-input-button-icon trailing edge"
+      icon={trailingButtonIcon}
+      type="transparent"
+      onClick={() => onClickTrailingButton()}
+      {...trailingButtonProps}
+    >
+      {trailingButtonText}
+    </Button>
+  );
+
   return (
     <div className={classNames('rc-input-container', className)} style={style}>
-      {icon && (
-        <Icon
-          className="rc-input-icon leading"
-          width="16px"
-          height="16px"
-          type={icon}
-        />
-      )}
-      {trailingIcon && (
-        <Icon
-          className="rc-input-icon trailing"
-          width="16px"
-          height="16px"
-          type={trailingIcon}
-        />
-      )}
+      {icon && lIcon}
+      {trailingIcon && tIcon}
+      {showTrailingButton && trailingButton}
       <Element
         id={name}
         name={name}